feat(供应链3.2):fix bug

+++ health-mall-product-model/src/main/java/com/drstrong/health/product/model/dto/product/PackageSkuDetailDTO.java
@@ -43,13 +43,13 @@ public class PackageSkuDetailDTO implements Serializable {
      * 原价
      */
     @ApiModelProperty("原价")
-    private Long originalPrice;
+    private BigDecimal originalPrice;
 
     /**
      * 优惠价
      */
     @ApiModelProperty(" 优惠价")
-    private Long preferentialPrice;
+    private BigDecimal preferentialPrice;
 
     /**
      * sku数量
